Bot controller: added middlewares
